Move panel buttons into subdirectory in panels
