+++ client/galaxy/scripts/components/Panels/Buttons/FavoritesButton.vue
