+++ client/galaxy/scripts/components/Panels/Buttons/UploadButton.vue
