+++ client/galaxy/scripts/components/Panels/Tool.vue
