+++ client/galaxy/scripts/components/Panels/ToolSearch.vue
