+++ client/galaxy/scripts/components/Panels/ToolSection.vue
